+++ framework/validator/__test__/typebox/typeEquality.test.ts
@@ -178,7 +178,7 @@ const shortTwo = schemify({
 type ShortExpected = {
   s: string;
   non: number;
-  b: Uint8Array;
+  b: Uint8Array<ArrayBuffer>;
   f: Blob;
 };